Add support for integer division
Set the compiler define to use integer division.

As per the limitations.md document, mRuby natively returns float results
from integer division (e.g. 1 / 2 => 0.5). This is a problem for my
application, where much of the Ruby code implicitly depends on integer
results from integer division.

There is a pull request on the mRuby project to add compile-time
configurable support for integer division, but it hasn't been merged in
over a year and may not be due to some disagreement over the issue. See
the link below.

I've applied the change 'by hand' which effectively separates the
submodule from the tip of the stable mRuby. Future updates from mRuby
will have to have this change applied for compatibility, until such time
as the change (or an equivalent) is supported in mainline mRuby.

https://github.com/mruby/mruby/pull/3123